Extract `getComparisonTarget` function
