+++ src/App.js
@@ -7,16 +7,9 @@ import { ChevronLeft, ChevronRight, Rewind } from "@core-ds/icons/16";
 import LinearProgress from "@material-ui/core/LinearProgress";
 import Button from "@material-ui/core/Button";
 import TextField from "@material-ui/core/TextField";
+import { getComparisonTarget } from "./comparison_urls.js";
 
 const { pages, comparison_target, default_comparison_source } = config;
-function getComparisonTarget(domain) {
-  const parts = domain.split(".");
-  if (parts.length > 3) {
-    const subdomain = parts.slice(0, -3).join(".");
-    return `${subdomain}.${comparison_target}`;
-  }
-  return comparison_target;
-}
 
 const Controls = styled.div`
   display: flex;
